+++ alicloud/resource_alicloud_db_account.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/denverdino/aliyungo/rds"
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/rds"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
 )
@@ -42,7 +42,7 @@ func resourceAlicloudDBAccount() *schema.Resource {
 			"type": &schema.Schema{
 				Type:         schema.TypeString,
 				Optional:     true,
-				ValidateFunc: validateAllowedStringValue([]string{string(rds.Normal), string(rds.Super)}),
+				ValidateFunc: validateAllowedStringValue([]string{string(Normal), string(Super)}),
 				Default:      "Normal",
 			},
 
@@ -55,23 +55,23 @@ func resourceAlicloudDBAccount() *schema.Resource {
 }
 
 func resourceAlicloudDBAccountCreate(d *schema.ResourceData, meta interface{}) error {
-	conn := meta.(*AliyunClient).rdsconn
-	args := rds.CreateAccountArgs{
-		DBInstanceId:    d.Get("instance_id").(string),
-		AccountName:     d.Get("name").(string),
-		AccountPassword: d.Get("password").(string),
-		AccountType:     rds.AccountType(d.Get("type").(string)),
-	}
+	client := meta.(*AliyunClient)
+	request := rds.CreateCreateAccountRequest()
+	request.DBInstanceId = d.Get("instance_id").(string)
+	request.AccountName = d.Get("name").(string)
+	request.AccountPassword = d.Get("password").(string)
+	request.AccountType = d.Get("type").(string)
+
 	if v, ok := d.GetOk("description"); ok && v.(string) != "" {
-		args.AccountDescription = v.(string)
+		request.AccountDescription = v.(string)
 	}
 	// wait instance running before modifying
-	if err := conn.WaitForInstanceAsyn(args.DBInstanceId, rds.Running, 500); err != nil {
-		return fmt.Errorf("WaitForInstance %s got error: %#v", rds.Running, err)
+	if err := client.WaitForDBInstance(request.DBInstanceId, Running, 500); err != nil {
+		return fmt.Errorf("WaitForInstance %s got error: %#v", Running, err)
 	}
-	err := resource.Retry(3*time.Minute, func() *resource.RetryError {
-		ag := args
-		if _, err := conn.CreateAccount(&ag); err != nil {
+	err := resource.Retry(5*time.Minute, func() *resource.RetryError {
+		args := request
+		if _, err := client.rdsconn.CreateAccount(args); err != nil {
 			if IsExceptedError(err, InvalidAccountNameDuplicate) {
 				return resource.NonRetryableError(fmt.Errorf("The account %s has already existed. Please import it using ID '%s:%s' or specify a new 'name' and try again.",
 					args.AccountName, args.DBInstanceId, args.AccountName))
@@ -88,10 +88,10 @@ func resourceAlicloudDBAccountCreate(d *schema.ResourceData, meta interface{}) e
 		return err
 	}
 
-	d.SetId(fmt.Sprintf("%s%s%s", args.DBInstanceId, COLON_SEPARATED, args.AccountName))
+	d.SetId(fmt.Sprintf("%s%s%s", request.DBInstanceId, COLON_SEPARATED, request.AccountName))
 
-	if err := meta.(*AliyunClient).rdsconn.WaitForAccount(args.DBInstanceId, args.AccountName, rds.Available, 500); err != nil {
-		return fmt.Errorf("Wait db account %s got an error: %#v.", rds.Available, err)
+	if err := client.WaitForAccount(request.DBInstanceId, request.AccountName, Available, 500); err != nil {
+		return fmt.Errorf("Wait db account %s got an error: %#v.", Available, err)
 	}
 
 	return resourceAlicloudDBAccountUpdate(d, meta)
@@ -102,7 +102,7 @@ func resourceAlicloudDBAccountRead(d *schema.ResourceData, meta interface{}) err
 	parts := strings.Split(d.Id(), COLON_SEPARATED)
 	account, err := meta.(*AliyunClient).DescribeDatabaseAccount(parts[0], parts[1])
 	if err != nil {
-		if NotFoundError(err) || IsExceptedError(err, InvalidDBInstanceIdNotFound) || IsExceptedError(err, InvalidAccountNameNotFound) {
+		if NotFoundDBInstance(err) {
 			d.SetId("")
 			return nil
 		}
@@ -126,18 +126,25 @@ func resourceAlicloudDBAccountUpdate(d *schema.ResourceData, meta interface{}) e
 
 	if d.HasChange("description") && !d.IsNewResource() {
 
-		if err := meta.(*AliyunClient).rdsconn.ModifyAccountDescription(&rds.ModifyAccountDescriptionArgs{
-			DBInstanceId:       instanceId,
-			AccountName:        accountName,
-			AccountDescription: d.Get("description").(string),
-		}); err != nil {
+		request := rds.CreateModifyAccountDescriptionRequest()
+		request.DBInstanceId = instanceId
+		request.AccountName = accountName
+		request.AccountDescription = d.Get("description").(string)
+
+		if _, err := meta.(*AliyunClient).rdsconn.ModifyAccountDescription(request); err != nil {
 			return fmt.Errorf("ModifyAccountDescription got an error: %#v", err)
 		}
 		d.SetPartial("description")
 	}
 
 	if d.HasChange("password") && !d.IsNewResource() {
-		if _, err := client.rdsconn.ResetAccountPassword(instanceId, accountName, d.Get("password").(string)); err != nil {
+
+		request := rds.CreateResetAccountPasswordRequest()
+		request.DBInstanceId = instanceId
+		request.AccountName = accountName
+		request.AccountPassword = d.Get("password").(string)
+
+		if _, err := client.rdsconn.ResetAccountPassword(request); err != nil {
 			return fmt.Errorf("Error reset db account password error: %#v", err)
 		}
 		d.SetPartial("password")
@@ -150,8 +157,12 @@ func resourceAlicloudDBAccountUpdate(d *schema.ResourceData, meta interface{}) e
 func resourceAlicloudDBAccountDelete(d *schema.ResourceData, meta interface{}) error {
 	parts := strings.Split(d.Id(), COLON_SEPARATED)
 
+	request := rds.CreateDeleteAccountRequest()
+	request.DBInstanceId = parts[0]
+	request.AccountName = parts[1]
+
 	return resource.Retry(5*time.Minute, func() *resource.RetryError {
-		if _, err := meta.(*AliyunClient).rdsconn.DeleteAccount(parts[0], parts[1]); err != nil {
+		if _, err := meta.(*AliyunClient).rdsconn.DeleteAccount(request); err != nil {
 			if IsExceptedError(err, InvalidAccountNameNotFound) {
 				return nil
 			}
@@ -160,7 +171,7 @@ func resourceAlicloudDBAccountDelete(d *schema.ResourceData, meta interface{}) e
 
 		resp, err := meta.(*AliyunClient).DescribeDatabaseAccount(parts[0], parts[1])
 		if err != nil {
-			if NotFoundError(err) || IsExceptedError(err, InvalidDBInstanceIdNotFound) || IsExceptedError(err, InvalidAccountNameNotFound) {
+			if NotFoundDBInstance(err) {
 				return nil
 			}
 			return resource.NonRetryableError(err)