+++ alicloud/resource_alicloud_slb.go
@@ -255,7 +255,7 @@ func resourceAliyunSlbCreate(d *schema.ResourceData, meta interface{}) error {
 
 	d.SetId(lb.LoadBalancerId)
 
-	if err := slbconn.WaitForLoadBalancerAsyn(lb.LoadBalancerId, slb.ActiveStatus, defaultTimeout); err != nil {
+	if err := slbconn.WaitForLoadBalancerAsyn(lb.LoadBalancerId, slb.ActiveStatus, DefaultTimeout); err != nil {
 		return fmt.Errorf("WaitForLoadbalancer %s got error: %#v", slb.ActiveStatus, err)
 	}