+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 IMPROVEMENTS:
 
 - Support golang version 1.9 ([#114](https://github.com/terraform-providers/terraform-provider-alicloud/pull/114))
+- RDS supports alibaba cloud official go sdk ([#113](https://github.com/terraform-providers/terraform-provider-alicloud/pull/113))
+
 
 
 ## 1.7.2 (February 09, 2018)